requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=1.7.0
torchvision>=0.8.0
pillow>=10.0.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.64.0
requests>=2.28.0
matplotlib>=3.3.0
seaborn>=0.11.0
scipy>=1.4.1
pyyaml>=5.3.1
gitpython>=3.1.30